Inputfield mit in AddToMovielist geschoben um State besser managen zu können
